refactoring

+++ src/main/java/senla/util/Paginator.java
@@ -2,20 +2,20 @@ package senla.util;
 
 public class Paginator {
 
-    public static Long getFirstElement(Long pageNumber, Long totalElements, Integer elementsPerPage) {
-        Long firstResult = (pageNumber - 1) * elementsPerPage;
+    public static Integer getFirstElement(Integer pageNumber, Long totalElements, Integer elementsPerPage) {
+        Integer firstResult = (pageNumber - 1) * elementsPerPage;
 
         if (firstResult < 0) {
-            firstResult = 0L;
+            firstResult = 0;
         } else if (firstResult > totalElements) {
-            firstResult = Paginator.getLastPageNumber(totalElements, elementsPerPage);
+            firstResult = Paginator.getLastPageNumber(Math.toIntExact(totalElements), elementsPerPage);
         }
 
         return firstResult;
     }
 
-    public static Long getLastPageNumber(Long totalElements, Integer elementsPerPage) {
-        Long remainder = totalElements % elementsPerPage;
+    public static Integer getLastPageNumber(Integer totalElements, Integer elementsPerPage) {
+        Integer remainder = totalElements % elementsPerPage;
         if (remainder == 0) {
             return totalElements - elementsPerPage;
         } else {
